requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.2
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.12.2
billiard==4.2.0
bitarray==2.8.2
blinker==1.6.2
black
boto3==1.28.55
botocore==1.31.55
bs4==0.0.1
cachetools==5.3.3
celery==5.4.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cytoolz==0.12.2
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
eth-abi==4.2.1
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.1
eth-utils==2.3.0
executing==1.2.0
filelock==3.12.4
flake8==6.0.0
Flask==2.3.2
Flask-Cors==3.0.10
flower==2.0.1
frozenlist==1.4.0
google-auth==2.6.0
gunicorn==21.2.0
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.5
httpx==0.27.0
humanize==4.8.0
idna==3.4
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kombu==5.3.7
lru-dict==1.2.0
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mongoengine==0.28.2
multidict==6.0.4
openai==1.19.0
packaging==23.1
parsimonious==0.9.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.17.1
prompt-toolkit==3.0.38
protobuf==4.24.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.19.0
pycurl==7.45.3
pydantic==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.8.0
pymongo==4.6.3
pymongo[srv]
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyunormalize==15.0.0
redis==4.5.4
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
rlp==3.0.0
rpds-py==0.10.6
rsa==4.9
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.2
tldextract==3.6.0
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.2
traitlets==5.9.0
typing_extensions==4.8.0
tzdata==2024.1
urllib3==1.26.16
uuid
vine==5.1.0
wcwidth==0.2.6
web3==6.11.1
webargs==8.2.0
websockets==12.0
Werkzeug==2.3.3
yarl==1.9.2
zipp==3.15.0
google-auth==2.6.0
pymongo[srv]
openai
uuid
keybert
sentence_transformers
presidio-analyzer     # Detects PII in text
presidio-anonymizer   # Anonymizes detected PII
imgurpython           # IMGUR uploading chart images
gunicorn
celery
nltk